+++ lib/kt_utils.dart
@@ -5,3 +5,5 @@ import 'dart:math';
 part 'string_kt.dart';
 
 part 'double_kt.dart';
+
+part 'list_kt.dart';