+++ CHANGELOG.md
@@ -1,7 +1,20 @@
-## 0.0.1
+## 0.1.0
 
-* Init version
+- Add:
+    - `reversed`
+    - `toDouble`
+    - `forEachIndexed`
+    - `lastIndex`
+    - `random`
+- Fix:
+    - Fix `roundPrecision` method
+- Update `README.md`
+
 
 ## 0.0.2
 
-* Change Description
+- Change Description
+
+## 0.0.1
+
+- Initial version